app/api/requirements.txt
Django==5.0.4
docker==7.0.0
djangorestframework==3.14.0
django-cors-headers
requests==2.31.0
